+++ src/Validation/ErrorList/TableEntriesSnapshot.cs
@@ -1,6 +1,4 @@
 ﻿using EnvDTE;
-using EnvDTE80;
-using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
 using Microsoft.VisualStudio.Shell.TableManager;
 using System.Collections.Generic;